Fixed issue when trying to commit multiple files at once
Also fixed issue where a commit message was being saved, thus being
committed even though the view was not active.  This was noticed when
using the build command, it saved all files before the build.  Now the
code checks to make sure the view we are given is in fact the active
view.