Fixed issue #3.  SVN setting now load after Sublime starts
Sublime was failing to load the settings file immediately upon
starting.  The SVNPlugin attempted to load the settingn right away
but Sublime was not ready.  To fix this issue, I set a boolean flag
to know whether setting had been loaded previously.  This check is
done when the caller of the sub-setting function actually calls the
settings.  So no settings are even loaded until a call into the
plugin is actually done.