Removed the initial line on the commit file
The code in the svn commit command that would highlight that first line of the commit log
no longer worked as of Sublime build 3124.  Don't know what changed to make it stop workings
so just removed that first line of code and the code that was used to highlight it.